Add shift hooks into individual files and send stockLineID with inserts
