+++ packages/inventory/src/Stocktake/DetailView/modal/StocktakeLineEdit/index.ts
@@ -1,1 +1,2 @@
 export * from './StocktakeLineEdit';
+export * from './utils';
