+++ packages/inventory/src/Stocktake/api/operations.graphql
@@ -23,6 +23,7 @@ fragment StocktakeLine on StocktakeLineNode {
   countedNumberOfPacks
   sellPricePerPack
   costPricePerPack
+
   location {
     __typename
     id
